Cleanup iOS and unbreak Darwin
