+++ tun_darwin.go
@@ -1,21 +1,17 @@
+// +build !ios
+
 package nebula
 
 import (
-	"errors"
 	"fmt"
-	"io"
+	"github.com/songgao/water"
 	"net"
-	"os"
 	"os/exec"
 	"strconv"
-	"sync"
-	"syscall"
-
-	"github.com/songgao/water"
 )
 
 type Tun struct {
-	io.ReadWriteCloser
+	// We have our own
 	Device       string
 	Cidr         *net.IPNet
 	MTU          int
@@ -24,32 +20,21 @@ type Tun struct {
 	*water.Interface
 }
 
-func newTunFromFd(deviceFd int, cidr *net.IPNet, defaultMTU int, routes []route, unsafeRoutes []route, txQueueLen int) (ifce *Tun, err error) {
-
-	file := os.NewFile(uintptr(deviceFd), "/dev/tun")
-
-	ifce = &Tun{
-		ReadWriteCloser: &tunReadCloser{f: file},
-		//fd:              int(file.Fd()),
-		Cidr: cidr,
-		MTU:  defaultMTU,
-		//TXQueueLen:      txQueueLen,
-		//Routes:          routes,
-		//UnsafeRoutes:    unsafeRoutes,
-	}
-	return
-}
-
 func newTun(deviceName string, cidr *net.IPNet, defaultMTU int, routes []route, unsafeRoutes []route, txQueueLen int) (ifce *Tun, err error) {
 	if len(routes) > 0 {
-		return nil, fmt.Errorf("Route MTU not supported in Darwin")
+		return nil, fmt.Errorf("route MTU not supported in Darwin")
 	}
+
 	// NOTE: You cannot set the deviceName under Darwin, so you must check tun.Device after calling .Activate()
 	return &Tun{
-		Cidr:         cidr,
-		MTU:          defaultMTU,
+		Cidr: cidr,
+		MTU:  defaultMTU,
 		UnsafeRoutes: unsafeRoutes,
 	}, nil
+}
+
+func newTunFromFd(deviceFd int, cidr *net.IPNet, defaultMTU int, routes []route, unsafeRoutes []route, txQueueLen int) (ifce *Tun, err error) {
+	return nil, fmt.Errorf("newTunFromFd not supported in Darwin")
 }
 
 func (c *Tun) Activate() error {
@@ -58,7 +43,7 @@ func (c *Tun) Activate() error {
 		DeviceType: water.TUN,
 	})
 	if err != nil {
-		return fmt.Errorf("Activate failed: %v", err)
+		return fmt.Errorf("activate failed: %v", err)
 	}
 
 	c.Device = c.Interface.Name()
@@ -86,66 +71,4 @@ func (c *Tun) Activate() error {
 func (c *Tun) WriteRaw(b []byte) error {
 	_, err := c.Write(b)
 	return err
-}
-
-// tunReadCloser is a hack to work around the first 4 bytes "packet
-// information" because there doesn't seem to be an IFF_NO_PI for darwin.
-type tunReadCloser struct {
-	f io.ReadWriteCloser
-
-	rMu  sync.Mutex
-	rBuf []byte
-
-	wMu  sync.Mutex
-	wBuf []byte
-}
-
-var _ io.ReadWriteCloser = (*tunReadCloser)(nil)
-
-func (t *tunReadCloser) Read(to []byte) (int, error) {
-	t.rMu.Lock()
-	defer t.rMu.Unlock()
-
-	if cap(t.rBuf) < len(to)+4 {
-		t.rBuf = make([]byte, len(to)+4)
-	}
-	t.rBuf = t.rBuf[:len(to)+4]
-
-	n, err := t.f.Read(t.rBuf)
-	copy(to, t.rBuf[4:])
-	return n - 4, err
-}
-
-func (t *tunReadCloser) Write(from []byte) (int, error) {
-
-	if len(from) == 0 {
-		return 0, syscall.EIO
-	}
-
-	t.wMu.Lock()
-	defer t.wMu.Unlock()
-
-	if cap(t.wBuf) < len(from)+4 {
-		t.wBuf = make([]byte, len(from)+4)
-	}
-	t.wBuf = t.wBuf[:len(from)+4]
-
-	// Determine the IP Family for the NULL L2 Header
-	ipVer := from[0] >> 4
-	if ipVer == 4 {
-		t.wBuf[3] = syscall.AF_INET
-	} else if ipVer == 6 {
-		t.wBuf[3] = syscall.AF_INET6
-	} else {
-		return 0, errors.New("Unable to determine IP version from packet")
-	}
-
-	copy(t.wBuf[4:], from)
-
-	n, err := t.f.Write(t.wBuf)
-	return n - 4, err
-}
-
-func (t *tunReadCloser) Close() error {
-	return t.f.Close()
 }
